Refactor data loading logic
Uses top-level exports in each respective data folder to abstract
implementation details, allowing for reduced coupling and easier
modifications to the data source should it be necessary in the future.
