+++ src/types/data.ts
@@ -1,0 +1,11 @@
+type ProjectLink = {
+  icon: string;
+  url: string;
+  label: string;
+};
+
+export type Project = {
+  name: string;
+  description?: string;
+  links: ProjectLink[];
+};
